Change DOS_File flags from uint32_t to uint8_t
The flags are a bitfield specifically of 8 bits.
They get set by an 8 bit register inside DOS interrupts.
Nothing ever uses or sets anything wider than 8 bits.

+++ src/dos/drive_overlay.cpp
@@ -216,7 +216,7 @@ public:
 	}
 
 	bool Write(uint8_t * data,uint16_t * size) override {
-		uint32_t f = flags&0xf;
+		uint8_t f = flags&0xf;
 		if (!overlay_active && (f == OPEN_READWRITE || f == OPEN_WRITE)) {
 			if (logoverlay) LOG_MSG("write detected, switching file for %s",GetName());
 			if (*data == 0) {
